completed first of multiple integration tests for user story 1
